+++ lib/generators/active_record/templates/migration.rb
@@ -1,4 +1,4 @@
-class DeviseCreate<%= table_name.camelize %> < ActiveRecord::Migration
+class DeviseCreate<%= table_name.camelize %> < ActiveRecord::Migration<%= migration_version %>
   def change
     create_table :<%= table_name %> do |t|
 <%= migration_data -%>